Resolve pyChapel testing errors
These tests started failing in our nightly testing suite due to a long standing
bug that had been brought to the forefront by some clean up Michael did in the
Chapel compiler.  The bug is that the assignment operator for records that are
generic on multiple fields doesn't resolve, and the POD clean up looks for a
user-defined assignment overload (while determining if a type is POD or not)
by trying to resolve a call to it.  If the user hasn't defined an overload,
then the compiler runs into this error.  This problem goes away if the record
is only generic for one field, so I altered the type definition.