Fix config error on ubuntu machine
We switched to a different test machine because that machine had python 2.7
installed on it, but this meant that the configuration file's relative path
used to find the pych.json file wasn't accurate.

This patch doesn't complete resolve that problem, it just delays it while I
think of a better solution.  Should allow testing to pass again.