Mark all functions that modify an array's `blk` with the required pragma
https://github.com/chapel-lang/chapel/pull/4510 added an optimization to remove
an inner `blk` multiply for cases where the compiler could prove it wasn't
needed. However, it's kind of a lame optimization (see the PR for why) and it
requires marking all functions that modify an array's blk field. Pychapel is
messing around with array internals (which it shouldn't really be doing IMO)
so we need to mark the functions that modify blk.