Remove use of deprecated noRefCount flag from Chapel compilation
This flag was removed in Chapel PR 4912, as a clean up from it no longer being
useful.  It was also no longer needed in pyChapel, as the fix it was providing
had been performed in another manner after the array rework, so I removed the
flag and the comment referencing why it was used in the first place.

Verified that one of the failing tests now passes.  Travis can handle full
testing.