Add an error message for trying to send sparsely packed numpy arrays to Chapel
With the arrays view work, we ignore the blk field, as no Chapel arrays will
normally be created with it having meaning.  However, in PyChapel we create
unusual arrays - numpy can give us an array to translate which is not densely
packed and without the blk field, we have no way of understanding that array.

For now, add an error message to pyChapel when a user tries to send Chapel such
an array.  Next commit, I will lock this in with some tests and update the
failing case that motivated this investigation.