docs: Make line-limited literalincludes less fragile
Particularly, harden them against code motion.

Use :end-before: instead of line numbers to stop the full program
listings before the test routines.  To this end, add a sentinel
comment line to the python programs where the literalinclude should
end.

Use the :pyobject: line-selection helper in the one case it applies --
the pure python quant().  It can't be used for the Chapel version
because it won't include the leading @Chapel() with our current pinned
version of sphinx 1.3.6; fixed by https://github.com/sphinx-doc/sphinx/issues/3348

The test_finance_chapel_numpy.py file that includes the @Chapel()
quant is not included as a full file anywhere, only for this one
function and an import line, so use sentinel lines for those too.